Introduce move semantics, with helper to convert legacy code
Introduces move semantics when creating UniValue or adding data with
push_back or pushKV. In many cases using these methods can dramatically
reduce the amount of useless work done. To get the most benefit of these
changes it is necessary to walk through the code and update it by adding
e.g. std::move where appropriate.

To make this task easier it is possible to add

   #define WITH_UNIVALUE_COPY_OPERATIONS 0

before including univalue.h which disables all UniValue methods that
produce a copy. Note that the copy constructor still needs to be there
for use in e.g. std::vector, but it is now marked as explicit.

With WITH_UNIVALUE_COPY_OPERATIONS 0, the compiler will produce an
error in each case where a copy would have occurred. To fix these compile
errors, ideally use std::move. When a copy is still needed, the compile
error can be silenced by explicitly calling the copy() method of
UniValue.

In my experiment, when adding `#define WITH_UNIVALUE_COPY_OPERATIONS 0`
to Bitcoin's core_write.cpp and blockchain.cpp, the compile errors can be
fixed with 29 std::move(), 9 copy(), and 1 const&. This almost doubles
the performance of the BlockToJsonVerbose benchmark.